import ValidationError from wtforms and then create two methods.
validate_email takes in the data field and checks our database to confirm there is no user registered with that email address. If a user with the same email address is found,  a ValidationError is raised and the error message passed in is displayed. As a result, the form is not submitted.

validate_username checks to see if the username is unique and raises a ValidationError if another user with a similar username is found.

When a method inside a form class begins with validate_ it is considered as a validator and is run with the other validators for that input field.