first create sender_email where we store our email address.
We then create a mail_message function that takes in 4 parameters which are the subject of the email, the template which is where we create the message body, We pass in the template without an extension because we need to create the text version and a HTML version, the recipient and any keyword arguments.

We then create a Message instance and pass in the subject, the sender_email and the recipient.

We then set up the email body and HTML. We use the send method of the mailinstance and pass in the email instance.